root is now a package, cli moved to cli/blox
