+++ .gitignore
@@ -8,3 +8,4 @@
 /repository.yaml
 /blox
 /_build/
+cli/blox/blox
